Refactored helpers from mouseDown to seperate file. Moved moveCallback into baseClass
