+++ src/eventDispatchers/mouseEventHandlers/mouseDown.js
@@ -2,12 +2,15 @@ import external from './../../externalModules.js';
 // State
 import { getters, state } from './../../store/index.js';
 import { getToolState } from './../../stateManagement/toolState.js';
-// Manipulators
-import getHandleNearImagePoint from './../../manipulators/getHandleNearImagePoint.js';
-import moveAllHandles from './../../manipulators/moveAllHandles.js';
-import moveHandle from './../../manipulators/moveHandle.js';
 // Util
 import isMouseButtonEnabled from './../../util/isMouseButtonEnabled.js';
+import {
+  getToolsWithMovableHandles,
+  findAndMoveHandleNearImagePoint,
+  findHandleNearImagePoint,
+  findAndMoveAnnotationNearClick,
+  findAnnotationNearClick
+} from './annotationToolHelpers.js';
 // Todo: Where should these live?
 import getInteractiveToolsForElement from './../../store/getInteractiveToolsForElement.js';
 import getToolsWithDataForElement from './../../store/getToolsWithDataForElement.js';
@@ -83,7 +86,9 @@ export default function (evt) {
 
       firstToolWithMoveableHandles.handleSelectedCallback(evt, handle);
 
-      preventPropagation(evt);
+      evt.stopImmediatePropagation();
+      evt.stopPropagation();
+      evt.preventDefault();
 
     } else {
       findAndMoveHandleNearImagePoint(
@@ -132,7 +137,9 @@ export default function (evt) {
 
       firstToolWithPointNearClick.toolSelectedCallback(evt, toolData);
 
-      preventPropagation(evt);
+      evt.stopImmediatePropagation();
+      evt.stopPropagation();
+      evt.preventDefault();
     } else {
       findAndMoveAnnotationNearClick(
         element,
@@ -145,139 +152,4 @@ export default function (evt) {
 
     return;
   }
-}
-
-const getToolsWithMovableHandles = function (element, tools, coords) {
-  return tools.filter((tool) => {
-    const toolState = getToolState(element, tool.name);
-
-    for (let i = 0; i < toolState.data.length; i++) {
-      if (
-        getHandleNearImagePoint(
-          element,
-          toolState.data[i].handles,
-          coords,
-          state.clickProximity
-        ) !== undefined
-      ) {
-        return true;
-      }
-    }
-
-    return false;
-  });
-};
-
-export const findAndMoveHandleNearImagePoint = function (
-  element,
-  evt,
-  toolState,
-  toolName,
-  coords
-) {
-  for (let i = 0; i < toolState.data.length; i++) {
-    const data = toolState.data[i];
-    const handle = getHandleNearImagePoint(
-      element,
-      data.handles,
-      coords,
-      state.clickProximity
-    );
-
-    if (handle) {
-      // Todo: We've grabbed a handle, stop listening/ignore for MOUSE_MOVE
-      data.active = true;
-      moveHandle(
-        evt.detail,
-        toolName,
-        data,
-        handle,
-        () => {
-          data.active = false;
-        },
-        true // PreventHandleOutsideImage
-      );
-
-      preventPropagation(evt);
-
-      return;
-    }
-  }
-};
-
-const findHandleNearImagePoint = function (
-  element,
-  evt,
-  toolState,
-  toolName,
-  coords
-) {
-  for (let i = 0; i < toolState.data.length; i++) {
-    const data = toolState.data[i];
-    const handle = getHandleNearImagePoint(
-      element,
-      data.handles,
-      coords,
-      state.clickProximity
-    );
-
-    if (handle) {
-      return handle;
-    }
-  }
-};
-
-const findAndMoveAnnotationNearClick = function (
-  element,
-  evt,
-  toolState,
-  tool,
-  coords
-) {
-  const opt = tool.options || {
-    deleteIfHandleOutsideImage: true,
-    preventHandleOutsideImage: false
-  };
-
-  for (let i = 0; i < toolState.data.length; i++) {
-    const data = toolState.data[i];
-    const isNearPoint = tool.pointNearTool(element, data, coords);
-
-    if (isNearPoint) {
-      data.active = true;
-      // TODO: Ignore MOUSE_MOVE for a bit
-      // TODO: Why do this and `moveHandle` expose this in different
-      // TODO: ways? PreventHandleOutsideImage
-      moveAllHandles(evt, toolState.data[i], toolState, tool.name, opt, () => {
-        data.active = false;
-      });
-
-      preventPropagation(evt);
-
-      return;
-    }
-  }
-};
-
-const findAnnotationNearClick = function (
-  element,
-  evt,
-  toolState,
-  tool,
-  coords
-) {
-  for (let i = 0; i < toolState.data.length; i++) {
-    const data = toolState.data[i];
-    const isNearPoint = tool.pointNearTool(element, data, coords);
-
-    if (isNearPoint) {
-      return data;
-    }
-  }
-};
-
-function preventPropagation (evt) {
-  evt.stopImmediatePropagation();
-  evt.stopPropagation();
-  evt.preventDefault();
 }
